+++ crates/icy_board_engine/src/icy_board/state/ppl_session.rs
@@ -122,8 +122,8 @@ impl UserDataValue for PplSession {
         Ok(value)
     }
 
-    async fn set_property_value(&self, _vm: &mut crate::vm::VirtualMachine<'_>, _name: &unicase::Ascii<String>, _val: VariableValue) -> crate::Res<()> {
-        Err("SESSION properties are read-only".into())
+    async fn set_property_value(&self, _vm: &mut crate::vm::VirtualMachine<'_>, name: &unicase::Ascii<String>, _val: VariableValue) -> crate::Res<()> {
+        Err(format!("SESSION property {name} is read-only").into())
     }
 
     async fn call_function(