+++ crates/icy_board_engine/src/icy_board/state/ppl_terminal_info.rs
@@ -171,8 +171,8 @@ impl UserDataValue for PplTerminalInfo {
         Ok(value)
     }
 
-    async fn set_property_value(&self, _vm: &mut crate::vm::VirtualMachine<'_>, _name: &unicase::Ascii<String>, _val: VariableValue) -> crate::Res<()> {
-        Ok(())
+    async fn set_property_value(&self, _vm: &mut crate::vm::VirtualMachine<'_>, name: &unicase::Ascii<String>, _val: VariableValue) -> crate::Res<()> {
+        Err(format!("TERMINFO property {name} is read-only").into())
     }
 
     async fn call_function(